requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jwt==1.3.1
MarkupSafe==2.1.5
mysqlclient==2.2.4
nodeenv==1.8.0
prisma==0.13.1
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyMySQL==1.1.0
python-decouple==3.8
python-dotenv==1.0.1
setuptools==69.5.1
sniffio==1.3.1
SQLAlchemy==2.0.30
tomlkit==0.12.5
typing_extensions==4.11.0
Werkzeug==3.0.3
